refactor: redesign issues page loading logic and server/client components
