+++ src/app/issue/layout.tsx
@@ -1,0 +1,10 @@
+import { Suspense } from "react";
+
+
+export default function Layout({ children }: { children: React.ReactNode }) {
+    return <div className="p-8 w-full h-full">
+        <Suspense fallback={null}>
+            {children}
+        </Suspense>
+    </div>
+}
