+++ src/app/issue/page.tsx
@@ -1,533 +1,39 @@
-import { prisma } from '@/lib/prisma';
-import { IssuePage, Issue, PropertyValue, PropertyDefinition } from '../../components/issue/IssuePage';
-import { FilterCondition } from '@/lib/property/types';
-import { Prisma } from '@prisma/client';
 import { auth } from '@clerk/nextjs/server';
-import { SystemPropertyId } from '@/lib/property/constants';
-import { deserializeSort, SortConfig } from '@/lib/sort-utils';
-import { deserializeFilters } from '@/lib/filter';
+import { IssueTable } from '@/components/issue/IssueTable';
+import { loadSearchParams } from '@/lib/issue/validation';
+import { buildOrderByParams, getPropertyDefinitions, getIssues } from '@/lib/issue/query';
 
-// 页面主组件
-export default async function Page({ searchParams }: { searchParams: Promise<{ filters?: string, page?: string, perPage?: string, sort?: string }> }) {
-    try {
-        // 从 URL 参数获取筛选条件和分页信息
-        const resolvedParams = await searchParams;
+export default async function Page({ searchParams }: {
+    searchParams: Promise<{ filters?: string, page?: string, perPage?: string, sort?: string }>
+}) {
+    const { page, perPage, sort, filters } = await loadSearchParams(searchParams);
 
-        const filtersStr = resolvedParams.filters;
-        const filters = filtersStr ? deserializeFilters(filtersStr) : [];
+    const { orgId } = await auth();
+    if (!orgId) {
+        // TODO throw error
+        console.warn('没有获取到组织ID，无法查询工单数据');
+        return <div className='flex flex-col items-center justify-center h-screen'>
+            <span className='text-2xl font-bold text-gray-500'>You are not in any workspace, please join a workspace first.</span>
+        </div>
+    }
 
-        // 获取排序参数
-        const sortStr = resolvedParams.sort;
-        const sortConfigs = sortStr ? deserializeSort(sortStr) : [];
+    const orderBy = buildOrderByParams(sort || []);
+    const [propertyDefinitions, issuesResult] = await Promise.all([
+        getPropertyDefinitions(),
+        getIssues(filters, orgId.toString(), page, perPage, orderBy)
+    ]);
+    const totalPages = Math.ceil(issuesResult.total / perPage);
 
-        // 构建排序参数
-        const orderBy = buildOrderByParams(sortConfigs);
+    if (propertyDefinitions.length === 0) {
+        // TODO throw error
+        return <div className="p-8 text-red-500">Error: No property definitions found</div>;
+    }
 
-        // 获取分页参数
-        const pageParam = resolvedParams.page;
-        const perPageParam = resolvedParams.perPage;
-
-        // 解析分页参数，默认第1页，每页10条
-        const page = pageParam ? parseInt(pageParam, 10) : 1;
-        const perPage = perPageParam ? parseInt(perPageParam, 10) : 10;
-
-        const { orgId } = await auth();
-
-        // 必须有组织ID才能查询数据
-        if (!orgId) {
-            console.warn('没有获取到组织ID，无法查询工单数据');
-            return <div className='flex flex-col items-center justify-center h-screen'>
-                <span className='text-2xl font-bold text-gray-500'>You are not in any workspace, please join a workspace first.</span>
-            </div>
-        }
-
-        // 获取属性定义和 issue 数据
-        const [propertyDefinitions, issuesResult] = await Promise.all([
-            getPropertyDefinitions(),
-            getIssues(filters, orgId.toString(), page, perPage, orderBy)
-        ]);
-
-        // 计算总页数
-        const totalPages = Math.ceil(issuesResult.total / perPage);
-
-        // 检查是否获取到了属性定义
-        if (propertyDefinitions.length === 0) {
-            return <div className="p-8 text-red-500">Error: No property definitions found</div>;
-        }
-
-        // 渲染客户端组件，传递数据
-        return <IssuePage
+    return <>
+        <IssueTable
             issues={issuesResult.issues}
             propertyDefinitions={propertyDefinitions}
             pageCount={totalPages}
-        />;
-    } catch (error) {
-        // 捕获并显示错误
-        console.error('渲染页面时出错:', error);
-        return (
-            <div className="p-8 text-red-500">
-                <h1 className="text-2xl font-bold mb-4">Error</h1>
-                <p>{error instanceof Error ? error.message : 'Unknown error'}</p>
-            </div>
-        );
-    }
-}
-
-// 从数据库中获取属性定义
-async function getPropertyDefinitions(): Promise<PropertyDefinition[]> {
-    try {
-        // 查询所有属性定义
-        const properties = await prisma.property.findMany({
-            where: {
-                deletedAt: null
-            },
-            select: {
-                id: true,
-                name: true,
-                type: true,
-                config: true
-            }
-        });
-
-        // 转换为前端需要的格式
-        const result = properties.map(prop => ({
-            id: prop.id,
-            name: prop.name,
-            type: prop.type,
-            config: prop.config as Record<string, unknown> | undefined
-        }));
-
-        // 定义属性的优先级顺序
-        // TODO tech dept shadcn table 暂不支持配置 column 顺序，手动指定顺序
-        const priorityOrder: Record<string, number> = {
-            [SystemPropertyId.ID]: 1,        // 工单 ID
-            [SystemPropertyId.TITLE]: 2,     // 工单标题
-            [SystemPropertyId.STATUS]: 3,    // 工单状态
-            [SystemPropertyId.PRIORITY]: 4,  // 工单优先级
-            [SystemPropertyId.CATEGORY]: 5,  // 工单类别
-            [SystemPropertyId.DIAGNOSIS]: 6, // 工单诊断
-            [SystemPropertyId.MINERS]: 7,    // 矿机列表
-            [SystemPropertyId.ASIGNEE]: 8,   // 负责人
-            [SystemPropertyId.REPORTER]: 9,  // 报告人
-            [SystemPropertyId.CREATED_AT]: 10, // 创建时间
-            [SystemPropertyId.UPDATED_AT]: 11, // 更新时间
-        };
-
-        // 根据优先级顺序排序
-        result.sort((a, b) => {
-            const priorityA = priorityOrder[a.id] || 999; // 没有指定优先级的放到最后
-            const priorityB = priorityOrder[b.id] || 999;
-            return priorityA - priorityB;
-        });
-
-        return result;
-    } catch (error) {
-        console.error('获取属性定义失败:', error);
-        throw new Error('获取属性定义失败');
-    }
-}
-
-
-// 不带筛选条件的工单查询
-async function getIssuesWithoutFilter(
-    baseWhere: Prisma.issueWhereInput,
-    page: number,
-    pageSize: number,
-    orderBy?: string | undefined
-): Promise<{ issues: Issue[], total: number }> {
-    // 查询符合条件的工单总数
-    const total = await prisma.issue.count({
-        where: baseWhere
-    });
-
-    // 计算分页偏移量
-    const skip = (page - 1) * pageSize;
-
-    // 查询所有工单
-    let issues: { id: string }[];
-
-    if (orderBy) {
-        // 使用原始SQL查询进行复杂排序
-        // 先生成基本条件
-        const whereConditions = [];
-
-        if (baseWhere.deletedAt === null) {
-            whereConditions.push(`issue."deletedAt" IS NULL`);
-        }
-
-        if (baseWhere.workspace_id) {
-            whereConditions.push(`issue."workspace_id" = '${baseWhere.workspace_id}'`);
-        }
-
-        const whereClause = whereConditions.length > 0
-            ? `WHERE ${whereConditions.join(' AND ')}`
-            : '';
-
-        // 执行原始SQL查询
-        const query = `
-            SELECT issue.id 
-            FROM issue 
-            ${whereClause}
-            ORDER BY ${orderBy}
-            LIMIT ${pageSize} OFFSET ${skip}
-        `;
-
-        issues = await prisma.$queryRawUnsafe<{ id: string }[]>(query);
-    } else {
-        // 使用默认排序
-        issues = await prisma.issue.findMany({
-            where: baseWhere,
-            select: {
-                id: true
-            },
-            orderBy: [{ createdAt: 'desc' }], // 默认按创建时间倒序排序
-            skip,
-            take: pageSize
-        });
-    }
-
-    if (issues.length === 0) {
-        return { issues: [], total };
-    }
-
-    // 查询属性值和构建返回数据
-    return await getIssuesWithValues(issues, total);
-}
-
-// 从数据库中获取工单数据
-async function getIssues(
-    filters: FilterCondition[] | undefined,
-    workspaceId: string,
-    page: number,
-    pageSize: number,
-    orderBy?: string | undefined
-): Promise<{ issues: Issue[], total: number }> {
-    try {
-        // 基本条件：未删除的工单并且必须指定 workspace_id
-        const baseWhere: Prisma.issueWhereInput = {
-            deletedAt: null,
-            workspace_id: workspaceId // 强制要求 workspace_id
-        };
-
-        // 如果有筛选条件，应用筛选
-        if (filters && filters.length > 0) {
-            // 由于 property_single_value 和 property_multi_value 不是 issue 的直接关系字段，
-            // 我们需要先获取符合条件的属性值记录，然后根据这些记录找到对应的 issue
-
-            // 对于每个筛选条件，构建查询
-            const issueIds = new Set<string>();
-
-            // 记录是否已经查询过
-            let hasSearched = false;
-
-            for (const filter of filters) {
-                // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                let valueCondition: any = {};
-
-                // 根据操作符构建查询条件
-                switch (filter.operator) {
-                    case 'contains':
-                        valueCondition = { contains: String(filter.value) };
-                        break;
-                    case 'eq':
-                        valueCondition = String(filter.value);
-                        break;
-                    case 'startsWith':
-                        valueCondition = { startsWith: String(filter.value) };
-                        break;
-                    case 'endsWith':
-                        valueCondition = { endsWith: String(filter.value) };
-                        break;
-                    case 'in':
-                        const values = Array.isArray(filter.value)
-                            ? filter.value.map(v => String(v))
-                            : [String(filter.value)];
-                        valueCondition = { in: values };
-                        break;
-                    default:
-                        throw new Error(`不支持的筛选操作符: ${filter.operator}`);
-                }
-
-                // 构建基本查询条件
-                const baseCondition = {
-                    property_id: filter.propertyId,
-                    property_type: filter.propertyType,
-                    value: valueCondition,
-                    deletedAt: null
-                };
-
-                // 根据属性类型确定查询表
-                let propertyValues: { issue_id: string }[] = [];
-
-                // TODO tech dept 这里做了个特殊判断，感觉是有问题的，回头排查
-                if (filter.propertyType === 'multi_select' || filter.propertyType === 'miners') {
-                    propertyValues = await prisma.property_multi_value.findMany({
-                        where: baseCondition,
-                        select: {
-                            issue_id: true
-                        },
-                        distinct: ['issue_id'] // 避免一个工单因为有多个匹配值而被重复计算
-                    });
-                } else {
-                    propertyValues = await prisma.property_single_value.findMany({
-                        where: baseCondition,
-                        select: {
-                            issue_id: true
-                        }
-                    });
-                }
-
-                // 如果是首个筛选条件，初始化结果集
-                if (!hasSearched) {
-                    propertyValues.forEach(pv => issueIds.add(pv.issue_id));
-                    hasSearched = true;
-                } else {
-                    // 如果已有结果集，取交集
-                    const currentIds = new Set<string>();
-                    propertyValues.forEach(pv => currentIds.add(pv.issue_id));
-
-                    // 保留在当前结果中也存在的ID
-                    for (const id of issueIds) {
-                        if (!currentIds.has(id)) {
-                            issueIds.delete(id);
-                        }
-                    }
-                }
-
-                // 如果已经没有符合条件的工单了，直接返回空结果
-                if (issueIds.size === 0 && hasSearched) {
-                    return { issues: [], total: 0 };
-                }
-            }
-
-            // 如果有筛选结果，查询这些工单
-            if (issueIds.size > 0) {
-                // 先获取符合筛选条件的总数量
-                const total = issueIds.size;
-
-                // 计算分页范围
-                const issueIdsArray = Array.from(issueIds);
-                const startIndex = (page - 1) * pageSize;
-                const endIndex = Math.min(startIndex + pageSize, issueIdsArray.length);
-                const paginatedIds = issueIdsArray.slice(startIndex, endIndex);
-
-                if (paginatedIds.length === 0) {
-                    return { issues: [], total };
-                }
-
-                // 构建 issue 查询条件
-                const whereCondition: Prisma.issueWhereInput = {
-                    ...baseWhere,
-                    id: { in: paginatedIds }
-                };
-
-                // 查询符合条件的工单
-                let issues: { id: string }[];
-
-                if (orderBy) {
-                    // 使用原始SQL查询进行复杂排序
-                    // 执行原始SQL查询，加上ID筛选条件
-                    const paginatedIdsFormatted = paginatedIds.map(id => `'${id}'`).join(',');
-
-                    if (!paginatedIdsFormatted) {
-                        return { issues: [], total };
-                    }
-
-                    const query = `
-                        SELECT issue.id 
-                        FROM issue 
-                        WHERE issue.id IN (${paginatedIdsFormatted})
-                        AND issue."workspace_id" = '${workspaceId}'
-                        AND issue."deletedAt" IS NULL
-                        ORDER BY ${orderBy}
-                        LIMIT ${pageSize}
-                    `;
-
-                    issues = await prisma.$queryRawUnsafe<{ id: string }[]>(query);
-                } else {
-                    // 使用默认排序
-                    issues = await prisma.issue.findMany({
-                        where: whereCondition,
-                        select: {
-                            id: true
-                        },
-                        orderBy: [{ createdAt: 'desc' }] // 默认按创建时间倒序排序
-                    });
-                }
-
-                if (issues.length === 0) {
-                    return { issues: [], total };
-                }
-
-                // 查询属性值和构建返回数据
-                return await getIssuesWithValues(issues, total);
-            } else {
-                // 没有筛选条件或没有找到符合条件的工单，查询所有工单
-                const result = await getIssuesWithoutFilter(baseWhere, page, pageSize, orderBy);
-                return result;
-            }
-        } else {
-            // 没有筛选条件，查询所有工单
-            const result = await getIssuesWithoutFilter(baseWhere, page, pageSize, orderBy);
-            return result;
-        }
-    } catch (error) {
-        console.error('获取工单数据失败:', error);
-        throw new Error('获取工单数据失败');
-    }
-}
-
-
-async function getIssuesWithValues(issues: { id: string }[], total: number): Promise<{ issues: Issue[], total: number }> {
-    // 批量查询所有工单的单值属性
-    const singlePropertyValues = await prisma.property_single_value.findMany({
-        where: {
-            issue_id: { in: issues.map(issue => issue.id) }, // 直接内联映射
-            deletedAt: null
-        },
-        select: {
-            issue_id: true,
-            property_id: true,
-            property_type: true,
-            value: true
-        }
-    });
-
-    // 批量查询所有工单的多值属性
-    const multiPropertyValues = await prisma.property_multi_value.findMany({
-        where: {
-            issue_id: { in: issues.map(issue => issue.id) }, // 直接内联映射
-            deletedAt: null
-        },
-        select: {
-            issue_id: true,
-            property_id: true,
-            property_type: true,
-            value: true,
-            position: true
-        },
-        orderBy: {
-            position: 'asc'
-        }
-    });
-
-    // 按工单ID分组数据
-    const singleValuesByIssueId = new Map<string, Array<typeof singlePropertyValues[0]>>();
-    const multiValuesByIssueId = new Map<string, Array<typeof multiPropertyValues[0]>>();
-
-    // 处理单值属性分组
-    for (const spv of singlePropertyValues) {
-        if (!singleValuesByIssueId.has(spv.issue_id)) {
-            singleValuesByIssueId.set(spv.issue_id, []);
-        }
-        singleValuesByIssueId.get(spv.issue_id)!.push(spv);
-    }
-
-    // 处理多值属性分组
-    for (const mpv of multiPropertyValues) {
-        if (!multiValuesByIssueId.has(mpv.issue_id)) {
-            multiValuesByIssueId.set(mpv.issue_id, []);
-        }
-        multiValuesByIssueId.get(mpv.issue_id)!.push(mpv);
-    }
-
-    // 构建结果集
-    const issueData: Issue[] = [];
-
-    for (const issue of issues) {
-        // 获取该工单的所有单值属性
-        const issueSingleValues = singleValuesByIssueId.get(issue.id) || [];
-
-        // 获取该工单的所有多值属性
-        const issueMultiValues = multiValuesByIssueId.get(issue.id) || [];
-
-        // 处理多值属性，将同一属性的多个值合并成数组
-        const multiValueMap = new Map<string, string[]>();
-
-        for (const mpv of issueMultiValues) {
-            if (!multiValueMap.has(mpv.property_id)) {
-                multiValueMap.set(mpv.property_id, []);
-            }
-            if (mpv.value) {
-                multiValueMap.get(mpv.property_id)?.push(mpv.value);
-            }
-        }
-
-        // 转换为前端需要的格式，先处理单值属性
-        const values: PropertyValue[] = issueSingleValues.map(pv => ({
-            property_id: pv.property_id,
-            value: pv.value
-        }));
-
-        // 添加多值属性
-        for (const [propertyId, valueArray] of multiValueMap.entries()) {
-            values.push({
-                property_id: propertyId,
-                value: valueArray
-            });
-        }
-
-        // 添加到工单数据中
-        issueData.push({
-            issue_id: issue.id,
-            property_values: values
-        });
-    }
-
-    return { issues: issueData, total };
-}
-
-
-/**
- * 将排序配置转换为Prisma原始SQL
- * 
- * 由于我们需要根据 property_single_value 或 property_multi_value 表中的字段进行排序，
- * 这里需要使用原始SQL查询来实现复杂的排序逻辑。
- * 
- * @param sortConfigs 排序配置数组
- * @returns 包含原始SQL的对象或undefined
- */
-function buildOrderByParams(sortConfigs: SortConfig[]): string | undefined {
-    if (!sortConfigs || sortConfigs.length === 0) {
-        return undefined; // 返回undefined表示使用默认排序
-    }
-
-    // 构建排序SQL表达式
-    const orderExpressions = sortConfigs.map(sort => {
-        const propertyId = sort.id;
-        const direction = sort.desc ? 'DESC' : 'ASC';
-
-        // 使用 SQL 子查询获取属性值，用于排序
-        // 如果是 ID 类型的属性，则优先使用 number_value 排序
-        if (propertyId === SystemPropertyId.ID) {
-            return `(
-                SELECT psv."number_value"
-                FROM property_single_value psv
-                WHERE psv."issue_id" = issue.id
-                AND psv."property_id" = '${propertyId}'
-                AND psv."deletedAt" IS NULL
-                LIMIT 1
-            ) ${direction} NULLS LAST`;
-        } else {
-            return `(
-                SELECT 
-                    CASE 
-                        WHEN psv."number_value" IS NOT NULL THEN psv."number_value"::text 
-                        ELSE psv.value 
-                    END
-                FROM property_single_value psv
-                WHERE psv."issue_id" = issue.id
-                AND psv."property_id" = '${propertyId}'
-                AND psv."deletedAt" IS NULL
-                LIMIT 1
-            ) ${direction} NULLS LAST`;
-        }
-    });
-
-    // 组合所有排序表达式
-    if (orderExpressions.length > 0) {
-        return orderExpressions.join(', ');
-    }
-
-    return undefined;
+        />
+    </>
 }
