--- src/app/page.tsx
@@ -1,7 +1,0 @@
-import { redirect } from 'next/navigation';
-
-const Home = () => {
-  redirect('/issue');
-}
-
-export default Home;
